server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cozmo==1.4.10
cozmoclad==3.4.0
distro==1.9.0
Flask==2.3.3
Flask-Cors==4.0.0
flask-helpers==0.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
openai==1.14.0
Pillow==10.0.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.0.4
Werkzeug==2.3.7
